# src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: "Medikm Apis Description"
  version: "1.0"
  title: "Medikm Apis"
...
schemes:
- "http"
- "https"
paths:
...
  /masterInput/addBiopsyType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Biopsy Type"
      description: "Biopsy examination type added."
      operationId: "addBiopsy"
      parameters:
      - in: "body"
        name: "body"
        description: "BiopsyTypeRequest"
        required: false
        schema:
          $ref: "#/definitions/Biopsy Type Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addCancerSubType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Cancer Sub Type"
      description: "Add Cancer Sub Type."
      operationId: "addCancerSubType"
      parameters:
      - in: "body"
        name: "body"
        description: "Cancer Sub Type Request"
        required: false
        schema:
          $ref: "#/definitions/Cancer SubType Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addCancerType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Cancer Type"
      description: "Add Cancer Type."
      operationId: "addCancerType"
      parameters:
      - in: "body"
        name: "body"
        description: "Cancer Type Request"
        required: false
        schema:
          $ref: "#/definitions/Cancer Type Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addFunctionalStatus:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Functional Status"
      description: "Add Functional Status."
      operationId: "addFunctionalStatus"
      parameters:
      - in: "body"
        name: "body"
        description: "Functional Status Request"
        required: false
        schema:
          $ref: "#/definitions/Functional Status Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addImaging:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Imaging"
      description: "Add Imaging."
      operationId: "addImaging"
      parameters:
      - in: "body"
        name: "body"
        description: "Imaging Status Request"
        required: false
        schema:
          $ref: "#/definitions/Imaging Status Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addPhysician:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Physician"
      description: "Add Physician."
      operationId: "addPhysician"
      parameters:
      - in: "body"
        name: "body"
        description: "Physician Request"
        required: false
        schema:
          $ref: "#/definitions/Physician Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Physician Response"
        302:
          description: "Exist | Username already exist."
          schema:
            $ref: "#/definitions/Physician Response"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/Physician Response"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/Physician Response"
  /masterInput/addRace/{raceName}:
    put:
      tags:
      - "Master Table Controller."
      summary: "Add Race"
      description: "Add Race"
      operationId: "addRace"
      parameters:
      - name: ""
        in: "path"
        description: "Race Name"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist | Username already exist."
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addReasonForPresentation:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Reason For Presentation"
      description: "Add Reason For Presentation."
      operationId: "addReasonForPresentation"
      parameters:
      - in: "body"
        name: "body"
        description: "Reason For Presentation Request"
        required: false
        schema:
          $ref: "#/definitions/Reason For Presentation Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/addTreatment:
    post:
      tags:
      - "Master Table Controller."
      summary: "Add Treatment"
      description: "Add Treatment"
      operationId: "addTreatment"
      parameters:
      - in: "body"
        name: "body"
        description: "Treatment Request"
        required: false
        schema:
          $ref: "#/definitions/Treatment Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editBiopsyType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Biopsy Type"
      description: "Biopsy examination type Edited."
      operationId: "editBiopsy"
      parameters:
      - in: "body"
        name: "body"
        description: "BiopsyTypeRequest"
        required: false
        schema:
          $ref: "#/definitions/Biopsy Type Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editCancerSubType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Cancer Sub Type"
      description: "Edit Cancer Sub Type."
      operationId: "editCancerSubType"
      parameters:
      - in: "body"
        name: "body"
        description: "Cancer Sub Type Request"
        required: false
        schema:
          $ref: "#/definitions/Cancer SubType Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editCancerType:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Cancer Type"
      description: "Edit Cancer Type."
      operationId: "editCancerType"
      parameters:
      - in: "body"
        name: "body"
        description: "Cancer Type Request"
        required: false
        schema:
          $ref: "#/definitions/Cancer Type Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editFunctionalStatus:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit FunctionalStatus"
      description: "Edit FunctionalStatus."
      operationId: "editFunctionalStatus"
      parameters:
      - in: "body"
        name: "body"
        description: "Functional Status Request"
        required: false
        schema:
          $ref: "#/definitions/Functional Status Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editImaging:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Imaging"
      description: "Edit Imaging."
      operationId: "editImaging"
      parameters:
      - in: "body"
        name: "body"
        description: "Imaging Status Request"
        required: false
        schema:
          $ref: "#/definitions/Imaging Status Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editPhysician:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Physician"
      description: "Edit Physician."
      operationId: "editPhysician"
      parameters:
      - in: "body"
        name: "body"
        description: "Physician Request"
        required: false
        schema:
          $ref: "#/definitions/Physician Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        302:
          description: "Exist | Username already exist."
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editReasonForPresentation:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Reason For Presentation"
      description: "Edit Reason For Presentation."
      operationId: "editReasonForPresentation"
      parameters:
      - in: "body"
        name: "body"
        description: "Reason For Presentation Request"
        required: false
        schema:
          $ref: "#/definitions/Reason For Presentation Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/editTreatment:
    post:
      tags:
      - "Master Table Controller."
      summary: "Edit Treatment"
      description: "Edit Treatment"
      operationId: "editTreatment"
      parameters:
      - in: "body"
        name: "body"
        description: "Treatment Request"
        required: false
        schema:
          $ref: "#/definitions/Treatment Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
  /masterInput/updateUserStatus:
    post:
      tags:
      - "Master Table Controller."
      summary: "Update User Status"
      description: "Update User Status"
      operationId: "updateUserStatus"
      parameters:
      - in: "body"
        name: "body"
        description: "Physician Request"
        required: false
        schema:
          $ref: "#/definitions/Physician Request"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/MedikmResponse"
        500:
          description: "Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
        525:
          description: "Authentication Failed"
          schema:
            $ref: "#/definitions/MedikmResponse"
...
definitions:
  Biopsy:
    type: "object"
    properties:
      lesionLocation:
        type: "string"
      needleGauge:
        type: "integer"
        format: "int32"
      postBiopsyRecommendedMRI:
        type: "integer"
        format: "int32"
      postBiopsyRelatedMRIControlateralBreast:
        type: "integer"
        format: "int32"
      lesionCompletelyRemoved:
        type: "integer"
        format: "int32"
      complications:
        type: "string"
      timeBiopsySampleTaken:
        type: "string"
        format: "date-time"
      timeBiopsySampleSent:
        type: "string"
        format: "date-time"
      biopsyMargins:
        type: "integer"
        format: "int32"
      biopsyMarginsComm:
        type: "string"
      lesionType:
        $ref: "#/definitions/LesionType"
      biopsyTypeId:
        $ref: "#/definitions/BiopsyType"
      caseId:
        $ref: "#/definitions/Case1"
      encounter:
        $ref: "#/definitions/Encounter"
      sninjMethod:
        type: "string"
...
  BiopsyType:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      biopsyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Biopsy"
      bname:
        type: "string"
...
  LesionType:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      biopsyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Biopsy"
      lname:
        type: "string"
...
  MedikmResponse:
    type: "object"
    properties:
      responseCode:
        type: "string"
        description: "Response Code"
      responseMessage:
        type: "string"
        description: "Response Message"
      authenticationKey:
        type: "string"
        description: "Authentication Key"
    description: "MedikmResponse class used for returning medikmresponse"
...
  Physician:
    type: "object"
    properties:
      physicianId:
        type: "integer"
        format: "int32"
      fname:
        type: "string"
      minit:
        type: "string"
      lname:
        type: "string"
      phyGroup:
        type: "string"
      phoneNo:
        type: "string"
      profileImage:
        type: "string"
      facilityId:
        $ref: "#/definitions/CareFacility"
      specialty:
        $ref: "#/definitions/PhysicianSpeciality"
      diseaseTypeId:
        $ref: "#/definitions/DiseaseType"
      cancerTypeId:
        $ref: "#/definitions/CancerType"
      caseResourcesCollection:
        type: "array"
        items:
          $ref: "#/definitions/CaseResources"
      clinicalTrialCollection:
        type: "array"
        items:
          $ref: "#/definitions/ClinicalTrial"
      casePhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/CasePhysician"
      encounterPhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/EncounterPhysician"
      case1Collection:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      case1Collection1:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      case1Collection2:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      case1Collection3:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      case1Collection4:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      userlogCollection:
        type: "array"
        items:
          $ref: "#/definitions/Userlog"
      tbDiscussionForumAnswereCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumAnswere"
      episodePhysicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/EpisodePhysician"
      pathologyCollection:
        type: "array"
        items:
          $ref: "#/definitions/Pathology"
      medikmResourcesCollection:
        type: "array"
        items:
          $ref: "#/definitions/MedikmResources"
      questionLikeCollection:
        type: "array"
        items:
          $ref: "#/definitions/QuestionLike"
      commentsLikeCollection:
        type: "array"
        items:
          $ref: "#/definitions/CommentsLike"
      tbDiscussionForumQuestionCollection:
        type: "array"
        items:
          $ref: "#/definitions/TbDiscussionForumQuestion"
      createdCaseCollection:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      lastUpdatedCaseCollection:
        type: "array"
        items:
          $ref: "#/definitions/Case1"
      queryBuilderCollection:
        type: "array"
        items:
          $ref: "#/definitions/QueryBuilder"
      emailId:
        type: "string"
...
  PhysicianSpeciality:
    type: "object"
    properties:
      typeId:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      physicianCollection:
        type: "array"
        items:
          $ref: "#/definitions/Physician"
...